requirements.txt
Django==4.0.5
django-countries[pyuca]==7.3.2
djangorestframework==3.13.1
Faker==13.13.0
psycopg2==2.9.3
PyJWT==2.1.0
python-decouple==3.6
Pillow==9.1.1